+++ Examples/Example-iOS_ObjC/Source/AppAuthExampleViewController.m
@@ -179,9 +179,10 @@ static NSString *const kAppAuthExampleAuthStateKey = @"authState";
                                                    grantTypes:nil
                                                   subjectType:nil
                                       tokenEndpointAuthMethod:@"client_secret_post"
-                                         additionalParameters:nil
-                                            additionalHeaders:nil];
-    // performs registration request
+                                           initialAccessToken:nil
+                                         additionalParameters:nil];
+
+  // performs registration request
     [self logMessage:@"Initiating registration request"];
 
     [OIDAuthorizationService performRegistrationRequest:request