+++ UnitTests/AppAuthTV/OIDTVAuthorizationRequestTests.m
@@ -21,10 +21,10 @@
 #if SWIFT_PACKAGE
 @import AppAuthTV;
 #else
-#import "Source/AppAuthCore/OIDScopeUtilities.h"
-#import "Source/AppAuthCore/OIDURLQueryComponent.h"
-#import "Source/AppAuthTV/OIDTVAuthorizationRequest.h"
-#import "Source/AppAuthTV/OIDTVServiceConfiguration.h"
+#import "Sources/AppAuthCore/OIDScopeUtilities.h"
+#import "Sources/AppAuthCore/OIDURLQueryComponent.h"
+#import "Sources/AppAuthTV/OIDTVAuthorizationRequest.h"
+#import "Sources/AppAuthTV/OIDTVServiceConfiguration.h"
 #endif
 
 // Ignore warnings about "Use of GNU statement expression extension" which is raised by our use of