+++ UnitTests/OIDGrantTypesTests.m
@@ -21,7 +21,7 @@
 #if SWIFT_PACKAGE
 @import AppAuthCore;
 #else
-#import "Source/AppAuthCore/OIDGrantTypes.h"
+#import "Sources/AppAuthCore/OIDGrantTypes.h"
 #endif
 
 // Ignore warnings about "Use of GNU statement expression extension" which is raised by our use of